+++ resources/bundle.js
@@ -1,0 +1,5 @@
+import registry from "@patternslib/patternslib/src/core/registry";
+
+import "./pat-tiles-management/tiles-management";
+
+registry.init();