+++ src/redturtle/tiles/management/event_handler.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from Acquisition import aq_base
 from persistent.list import PersistentList
 from persistent.mapping import PersistentMapping
@@ -7,9 +6,9 @@ import re
 
 
 def getManagerId(tile):
-    managerId = tile.request.form.get('managerId')
+    managerId = tile.request.form.get("managerId")
     if not managerId:
-        managerId = 'defaultManager'
+        managerId = "defaultManager"
     return managerId
 
 
@@ -23,19 +22,19 @@ def tileCreated(tile, event):
     managerId = getManagerId(tile)
 
     new_tile = PersistentMapping()
-    new_tile['tile_id'] = tile_id
-    new_tile['tile_hidden'] = False
-    new_tile['tile_style'] = ''
+    new_tile["tile_id"] = tile_id
+    new_tile["tile_hidden"] = False
+    new_tile["tile_style"] = ""
 
     try:
-        tile_type = re.search('@@(.*?)/', tile.url).group(1)
+        tile_type = re.search("@@(.*?)/", tile.url).group(1)
     except AttributeError:
-        tile_type = ''
+        tile_type = ""
     if tile_type:
-        new_tile['tile_type'] = tile_type
+        new_tile["tile_type"] = tile_type
 
     # store tiles_order in persistent object attribute.
-    if not getattr(context, 'tiles_list', {}):
+    if not getattr(context, "tiles_list", {}):
         context.tiles_list = PersistentMapping()
     if managerId not in context.tiles_list:
         context.tiles_list[managerId] = PersistentList()
@@ -45,12 +44,12 @@ def tileCreated(tile, event):
 def tileDeleted(tile, event):
     context = aq_base(tile.context)
     managerId = getManagerId(tile)
-    tiles_list = getattr(context, 'tiles_list', {})
+    tiles_list = getattr(context, "tiles_list", {})
     if not tiles_list:
         return
     tilesForManager = tiles_list.get(managerId, [])
     if not tilesForManager:
         return
     for tile_info in tilesForManager:
-        if tile_info.get('tile_id') == tile.id:
+        if tile_info.get("tile_id") == tile.id:
             context.tiles_list[managerId].remove(tile_info)