+++ src/redturtle/tiles/management/interfaces.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from redturtle.tiles.management import _
 from zope import schema
 from zope.interface import Interface
@@ -19,38 +18,38 @@ class ICreatedTile(Interface):
 
 class IRedturtleTilesManagementSettings(Interface):
     """ """
+
     enabled_tiles = schema.List(
-        title=_(u'enabled_tiles_label',
-                default=u'Enabled tiles'),
-        description=_(u'enabled_tiles_help',
-                      default=u'Select a list of tiles to add.'),
+        title=_("enabled_tiles_label", default="Enabled tiles"),
+        description=_("enabled_tiles_help", default="Select a list of tiles to add."),
         required=False,
         default=[],
         missing_value=[],
         value_type=schema.Choice(
-            vocabulary='tiles.management.vocabularies.RegisteredTiles',
+            vocabulary="tiles.management.vocabularies.RegisteredTiles",
         ),
     )
 
     tile_size_css_class = schema.List(
-        title=_(u'size_css_classes',
-                default=u'CSS Classes for tile sizes'),
-        description=_(u'size_css_classes_descriptions',
-                      default=u'List of CSS classes to resize the tile. '
-                              u'These are used in the size button in tile '
-                              u'management.\n'
-                              u'The default style is "reset". It will add an '
-                              u'empty string as CSS class and the tile will '
-                              u'take the whole width.\n'
-                              u'Insert a list of values (one per row) in the '
-                              u'following form: display_name|css_class where '
-                              u'display_name is the string to show to the user'
-                              u' and css_class is the class will be applied '
-                              u'to the tile'),
+        title=_("size_css_classes", default="CSS Classes for tile sizes"),
+        description=_(
+            "size_css_classes_descriptions",
+            default="List of CSS classes to resize the tile. "
+            "These are used in the size button in tile "
+            "management.\n"
+            'The default style is "reset". It will add an '
+            "empty string as CSS class and the tile will "
+            "take the whole width.\n"
+            "Insert a list of values (one per row) in the "
+            "following form: display_name|css_class where "
+            "display_name is the string to show to the user"
+            " and css_class is the class will be applied "
+            "to the tile",
+        ),
         required=False,
         default=[
-            'reset|',
-            'two tiles in a row|half-width',
+            "reset|",
+            "two tiles in a row|half-width",
         ],
         missing_value=[],
         value_type=schema.TextLine(),