+++ src/redturtle/tiles/management/testing.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from plone.app.contenttypes.testing import PLONE_APP_CONTENTTYPES_FIXTURE
 from plone.app.robotframework.testing import REMOTE_LIBRARY_BUNDLE_FIXTURE
 from plone.app.testing import applyProfile
@@ -11,14 +10,13 @@ import redturtle.tiles.management
 
 
 class RedturtleTilesManagementLayer(PloneSandboxLayer):
-
     defaultBases = (PLONE_APP_CONTENTTYPES_FIXTURE,)
 
     def setUpZope(self, app, configurationContext):
         self.loadZCML(package=redturtle.tiles.management)
 
     def setUpPloneSite(self, portal):
-        applyProfile(portal, 'redturtle.tiles.management:default')
+        applyProfile(portal, "redturtle.tiles.management:default")
 
 
 REDTURTLE_TILES_MANAGEMENT_FIXTURE = RedturtleTilesManagementLayer()
@@ -26,13 +24,13 @@ REDTURTLE_TILES_MANAGEMENT_FIXTURE = RedturtleTilesManagementLayer()
 
 REDTURTLE_TILES_MANAGEMENT_INTEGRATION_TESTING = IntegrationTesting(
     bases=(REDTURTLE_TILES_MANAGEMENT_FIXTURE,),
-    name='RedturtleTilesManagementLayer:IntegrationTesting',
+    name="RedturtleTilesManagementLayer:IntegrationTesting",
 )
 
 
 REDTURTLE_TILES_MANAGEMENT_FUNCTIONAL_TESTING = FunctionalTesting(
     bases=(REDTURTLE_TILES_MANAGEMENT_FIXTURE,),
-    name='RedturtleTilesManagementLayer:FunctionalTesting',
+    name="RedturtleTilesManagementLayer:FunctionalTesting",
 )
 
 
@@ -42,5 +40,5 @@ REDTURTLE_TILES_MANAGEMENT_ACCEPTANCE_TESTING = FunctionalTesting(
         REMOTE_LIBRARY_BUNDLE_FIXTURE,
         z2.ZSERVER_FIXTURE,
     ),
-    name='RedturtleTilesManagementLayer:AcceptanceTesting',
+    name="RedturtleTilesManagementLayer:AcceptanceTesting",
 )