+++ src/redturtle/tiles/management/tests/helpers.py
@@ -1,11 +1,9 @@
-# -*- coding: utf-8 -*-
 from plone.tiles.interfaces import ITile
 from zope.interface import implementer
 
 
 @implementer(ITile)
-class TestTile(object):
-
+class TestTile:
     def __init__(self, context, request, url, id):
         self.context = context
         self.request = request